1,修复界面显示出错的问题; 2,增加了一个数据库; 3,使用viewModel增加全局消息通信

+++ app/src/main/java/com/example/kotweather/common/util/AirUtil.kt
@@ -2,7 +2,7 @@ package com.example.kotweather.common.util
 
 class AirLevel(val airLevel: String)
 
-fun getAirLevel(airLevel: Int): AirLevel {
+fun getAirLevel(airLevel: Double): AirLevel {
     if (airLevel in 0.0..50.0) {
         return AirLevel("优")
     }
