+++ app/src/main/java/com/example/kotweather/model/HourlyData.kt
@@ -56,7 +56,7 @@ data class Humidity(
 
 data class Precipitation(
         val datetime: String,
-        val value: Int
+        val value: Double
 )
 
 data class Pressure(
@@ -92,10 +92,10 @@ data class Aqi(
 
 data class Pm25(
         val datetime: String,
-        val value: Int
+        val value: Double
 )
 
 data class Value(
-        val chn: Int,
-        val usa: Int
+        val chn: Double,
+        val usa: Double
 )
